+++ aoc2020/day1.h
@@ -8,7 +8,7 @@ using std::string;
 class day1
 {
 	public:
-		static int run(vector<string> input, bool partB);
+		static long long run(vector<string> input, bool partB);
 		static int partA(vector<int> input);
 		static int partB(vector<int> input);
 };
